refactor: split utils.ts into focused modules (#111)
