+++ src/content/tokens.ts
@@ -1,0 +1,26 @@
+export function splitIntoTokens(str: string): string[] {
+  // CJK ideographs → one token each (they carry meaning per character)
+  // Latin words/numbers → one token per word (whitespace-delimited)
+  // Punctuation (ASCII + CJK) → individual tokens (preserves structure)
+  // Newlines → tokens (headings and paragraphs depend on line breaks)
+  const regex = /[\u4e00-\u9fa5]|[a-zA-Z0-9]+|[.,!?;，。！？；#]|[\n]/g;
+  const tokens = str.match(regex);
+  return tokens || [];
+}
+
+export function joinTokens(tokens: string[]): string {
+  let result = "";
+  for (let i = 0; i < tokens.length; i++) {
+    const token = tokens[i];
+    if (token === "\n") {
+      result += token;
+    } else if (/[\u4e00-\u9fa5]|[.,!?;，。！？；#]/.test(token)) {
+      result += token;
+    } else {
+      const prevToken = i > 0 ? tokens[i - 1] : undefined;
+      const needsSpace = i > 0 && prevToken !== "\n";
+      result += (needsSpace ? " " : "") + token;
+    }
+  }
+  return result.trim();
+}
